ios: remove the dead iroh independent-event and artifact lanes
The iOS app never wires independentEventByteStreamProvider or
artifactLaneProvider (cmuxApp.swift constructs CMUXMobileRuntime without
them), and both factories were additionally gated on route.kind == .iroh,
which production supportedRouteKinds no longer contains. Deletes the
session's optional-lane negotiation (prepareIndependentServerEvents, the
event_transport advertisement, supportsPipelinedInitialEventSubscribe),
openArtifactLane and the chat artifact lane fetch path (which could only
mint a descriptor RPC and then always fall back), the lifecycle-gate
admissions for both lanes, the runtime slots, and the
irohDirectOnlyDialCandidates parameter no caller passed. dispatch(frame:)
moves to MobileCoreRPCSession+EventDispatch.swift because the control
reader still uses it. The CmxByteTransportRequest core field stays; the
Mac still hosts these lanes for old phones.
