+++ Packages/iOS/CmuxMobileRPC/Sources/CmuxMobileRPC/MobileCoreRPCSession.swift
@@ -3,7 +3,6 @@ import Foundation
 
 actor MobileCoreRPCSession {
     typealias TransportFactory = @Sendable () throws -> any CmxByteTransport
-    typealias IndependentEventByteStreamFactory = @Sendable () async throws -> CmxIndependentEventByteStream
     typealias ConnectedCandidateHook = @Sendable (_ candidate: any CmxByteTransport) async -> Void
     typealias TransportConnectObserver = @Sendable (MobileRPCTransportConnectEvent) -> Void
     typealias TearDownRegistrationHook = @Sendable () async -> Void
@@ -58,16 +57,6 @@ actor MobileCoreRPCSession {
         var cancelledRequestResolutionTask: Task<Void, Never>?
     }
 
-    struct IndependentEventPreparation: Sendable {
-        let id: UUID
-        let task: Task<CmxIndependentEventByteStream, any Error>
-    }
-
-    struct IndependentEventReader: Sendable {
-        let id: UUID
-        let task: Task<Void, Never>
-    }
-
     let taskTimeout = RPCTaskTimeout()
     private let connectAttemptKey: MobileRPCConnectAttemptKey?
     let connectAttemptRegistry: MobileRPCConnectAttemptRegistry
@@ -79,7 +68,6 @@ actor MobileCoreRPCSession {
     let lateAbandonedConnectCloseTimeoutNanoseconds: UInt64
     let cancelledWriteCompletionGraceNanoseconds: UInt64
     private let makeTransport: TransportFactory
-    let makeIndependentEventByteStream: IndependentEventByteStreamFactory?
     private let didReceiveConnectedCandidate: ConnectedCandidateHook?
     private let diagnosticTransport: DiagnosticTransportKind?
     private let transportConnectObserver: TransportConnectObserver?
@@ -100,11 +88,6 @@ actor MobileCoreRPCSession {
     private var recordedConnectCancellationAttemptIDs: Set<Int> = []
     private var installedConnectionID: UUID?
     private var readerTask: Task<Void, Never>?
-    var independentEventPreparation: IndependentEventPreparation?
-    var independentEventReader: IndependentEventReader?
-    /// Subscription stream IDs that already made their one optional-lane
-    /// negotiation attempt during this control-session generation.
-    var independentEventSubscriptionStreamIDs: Set<String> = []
     var pending: [String: PendingContinuation] = [:]
     var pipelinedPending: [String: PipelinedRequestSettlement] = [:]
     var requestTimeoutTasks: [String: Task<Void, Never>] = [:]
@@ -137,7 +120,6 @@ actor MobileCoreRPCSession {
         cancelledWriteCompletionGraceNanoseconds: UInt64 =
             MobileCoreRPCSession.defaultCancelledWriteCompletionGraceNanoseconds,
         makeTransport: @escaping TransportFactory,
-        makeIndependentEventByteStream: IndependentEventByteStreamFactory? = nil,
         didReceiveConnectedCandidate: ConnectedCandidateHook? = nil,
         diagnosticTransport: DiagnosticTransportKind? = nil,
         transportConnectObserver: TransportConnectObserver? = nil,
@@ -153,7 +135,6 @@ actor MobileCoreRPCSession {
         self.cancelledWriteCompletionGraceNanoseconds =
             cancelledWriteCompletionGraceNanoseconds
         self.makeTransport = makeTransport
-        self.makeIndependentEventByteStream = makeIndependentEventByteStream
         self.didReceiveConnectedCandidate = didReceiveConnectedCandidate
         self.diagnosticTransport = diagnosticTransport
         self.transportConnectObserver = transportConnectObserver
@@ -211,8 +192,6 @@ actor MobileCoreRPCSession {
             }
         }
         readerTask?.cancel()
-        independentEventPreparation?.task.cancel()
-        independentEventReader?.task.cancel()
         activeWrite?.task.cancel()
         activeWrite?.cancelledRequestResolutionTask?.cancel()
         writerTask?.cancel()
@@ -437,11 +416,6 @@ actor MobileCoreRPCSession {
         transport = nil
         readerTask?.cancel()
         readerTask = nil
-        independentEventPreparation?.task.cancel()
-        independentEventPreparation = nil
-        independentEventReader?.task.cancel()
-        independentEventReader = nil
-        independentEventSubscriptionStreamIDs.removeAll()
         await tearDownRegistrationHook?()
         if let transportToClose {
             await enqueueTransportClose(
