+++ Packages/iOS/CmuxMobileRPC/Tests/CmuxMobileRPCTests/TransportTestDoubles.swift
@@ -15,7 +15,6 @@ struct TestMobileSyncRuntime: MobileSyncRuntime {
     var pairingRequestTimeoutNanoseconds: UInt64
     var now: @Sendable () -> Date
     var supportsServerPushEvents: Bool
-    var independentEventByteStreamProvider: CmxIndependentEventByteStreamProvider?
 
     init(
         transportFactory: any CmxByteTransportFactory,
@@ -27,8 +26,7 @@ struct TestMobileSyncRuntime: MobileSyncRuntime {
         rpcRequestTimeoutNanoseconds: UInt64 = 30 * 1_000_000_000,
         pairingRequestTimeoutNanoseconds: UInt64 = 30 * 1_000_000_000,
         now: @escaping @Sendable () -> Date = Date.init,
-        supportsServerPushEvents: Bool = true,
-        independentEventByteStreamProvider: CmxIndependentEventByteStreamProvider? = nil
+        supportsServerPushEvents: Bool = true
     ) {
         self.supportedRouteKinds = supportedRouteKinds
         self.transportFactory = transportFactory
@@ -47,7 +45,6 @@ struct TestMobileSyncRuntime: MobileSyncRuntime {
         self.pairingRequestTimeoutNanoseconds = pairingRequestTimeoutNanoseconds
         self.now = now
         self.supportsServerPushEvents = supportsServerPushEvents
-        self.independentEventByteStreamProvider = independentEventByteStreamProvider
     }
 }
