+++ Packages/iOS/CmuxMobileShell/Sources/CmuxMobileShell/MobileChatEventSource.swift
@@ -35,8 +35,6 @@ public actor MobileChatEventSource: ChatEventSource {
     public nonisolated let supportsPanelArtifacts: Bool
     /// Whether the connected Mac supports session-wide artifact gallery pages.
     public nonisolated let supportsArtifactGallery: Bool
-    /// Whether raw artifact bytes may use a peer-bound Iroh application lane.
-    public nonisolated let supportsArtifactLane: Bool
 
     /// Creates the adapter.
     ///
@@ -48,7 +46,6 @@ public actor MobileChatEventSource: ChatEventSource {
         supportsArtifactFolders: Bool = false,
         supportsTerminalArtifactList: Bool = false,
         supportsPanelArtifacts: Bool = false,
-        supportsArtifactLane: Bool = false,
         diagnosticLog: DiagnosticLog? = nil
     ) {
         self.client = client
@@ -58,7 +55,6 @@ public actor MobileChatEventSource: ChatEventSource {
         self.supportsArtifactFolders = supportsArtifactFolders
         self.supportsTerminalArtifactList = supportsTerminalArtifactList
         self.supportsPanelArtifacts = supportsPanelArtifacts
-        self.supportsArtifactLane = supportsArtifactLane
     }
 
     /// Lists chat-capable agent sessions the Mac knows about.
@@ -495,57 +491,6 @@ public actor MobileChatEventSource: ChatEventSource {
         progress: (@Sendable (_ fetchedBytes: Int64, _ totalBytes: Int64) -> Void)?,
         onChunk: @Sendable (ChatArtifactChunk) async throws -> Void
     ) async throws -> Data {
-        if supportsArtifactLane {
-            let descriptor: ChatArtifactLaneDescriptor
-            let connection: any MobileArtifactLaneConnection
-            do {
-                var descriptorParams: [String: Any] = stringParams
-                descriptorParams["transport"] = "iroh_artifact_v1"
-                descriptor = try await artifactCall(
-                    method: method,
-                    params: descriptorParams
-                )
-                guard descriptor.totalSize >= 0 else {
-                    throw MobileArtifactLaneFetchError.invalidDescriptor
-                }
-                connection = try await client.openArtifactLane(
-                    resourceID: descriptor.resourceID,
-                    offset: 0
-                )
-            } catch is CancellationError {
-                throw CancellationError()
-            } catch {
-                // Descriptor mint/open failed before any lane byte was exposed.
-                // Preserve compatibility by using the existing RPC path.
-                return try await fetchArtifactChunksOverRPC(
-                    method: method,
-                    stringParams: stringParams,
-                    collectsData: collectsData,
-                    progress: progress,
-                    onChunk: onChunk
-                )
-            }
-            do {
-                return try await MobileArtifactLaneFetchLoop().run(
-                    descriptor: descriptor,
-                    connection: connection,
-                    collectsData: collectsData,
-                    progress: progress,
-                    onChunk: onChunk
-                )
-            } catch MobileArtifactLaneFetchError.failedBeforeFirstByte {
-                // No consumer-visible bytes were delivered, so the legacy
-                // authorized RPC can safely restart from offset zero.
-            } catch is CancellationError {
-                throw CancellationError()
-            } catch MobileArtifactLaneFetchError.failedAfterFirstByte {
-                // Once the lane exposed bytes, mixing in an RPC restart could
-                // splice two file versions into one preview.
-                throw ChatArtifactError.transferInterrupted
-            } catch MobileArtifactLaneFetchError.invalidDescriptor {
-                throw ChatArtifactError.invalidResponse
-            }
-        }
         return try await fetchArtifactChunksOverRPC(
             method: method,
             stringParams: stringParams,
