+++ Packages/iOS/CmuxMobileShell/Sources/CmuxMobileShell/MobileShellComposite+AgentChat.swift
@@ -15,7 +15,6 @@ extension MobileShellComposite {
             supportsArtifactFolders: supportsChatArtifactFolders,
             supportsTerminalArtifactList: supportsTerminalArtifactList,
             supportsPanelArtifacts: supportsPanelArtifacts,
-            supportsArtifactLane: supportsIrohArtifactLane,
             diagnosticLog: diagnosticLog
         )
     }
