+++ Packages/iOS/CmuxMobileShell/Tests/CmuxMobileShellTests/MobileShellWorkspaceCapabilityTests.swift
@@ -27,17 +27,14 @@ import Testing
         #expect(!store.supportsChatArtifactFolders)
         #expect(!store.supportsTerminalArtifactList)
         #expect(!store.supportsPanelArtifacts)
-        #expect(!store.supportsIrohArtifactLane)
 
         store.supportedHostCapabilities.formUnion([
             "chat.artifact.folders.v1",
             "terminal.artifact.list.v1",
-            "iroh.artifact_lane.v1",
             "panel.artifact.v1",
         ])
         #expect(store.supportsChatArtifactFolders)
         #expect(store.supportsTerminalArtifactList)
-        #expect(store.supportsIrohArtifactLane)
         #expect(store.supportsPanelArtifacts)
     }
