+++ ios/cmuxPackage/Sources/cmuxFeature/CMUXMobileRuntime.swift
@@ -8,9 +8,10 @@ import OSLog
 
 public struct CMUXMobileRuntime: Sendable, MobileSyncRuntime {
     public static let defaultRPCRequestTimeoutNanoseconds: UInt64 = 30 * 1_000_000_000
-    // Iroh may spend more than eight seconds establishing a relay path before
-    // the first authenticated RPC can be written. Keep one hard end-to-end
-    // pairing deadline, but size it to leave room for both phases.
+    // A pairing dial may spend several seconds establishing its transport
+    // (cold relay WebSocket, Tailscale path setup) before the first
+    // authenticated RPC can be written. Keep one hard end-to-end pairing
+    // deadline, but size it to leave room for both phases.
     public static let defaultPairingRequestTimeoutNanoseconds: UInt64 = 30 * 1_000_000_000
     public static let defaultPairingAttemptTimeoutNanoseconds: UInt64 = 30 * 1_000_000_000
 
@@ -33,10 +34,8 @@ public struct CMUXMobileRuntime: Sendable, MobileSyncRuntime {
     /// Production sets it on (the default), and falls back to the legacy
     /// 750ms poll only when a connected Mac does not support events.
     public var supportsServerPushEvents: Bool
-    public var independentEventByteStreamProvider: CmxIndependentEventByteStreamProvider?
     public var terminalLaneProvider: MobileTerminalLaneProvider?
     public var simulatorStreamLaneProvider: MobileSimulatorStreamLaneProvider?
-    public var artifactLaneProvider: MobileArtifactLaneProvider?
 
     /// Builds the production access-token provider over an injected
     /// ``TokenProviding`` (the app-root ``AuthCoordinator``), honoring the DEBUG
@@ -144,9 +143,7 @@ public struct CMUXMobileRuntime: Sendable, MobileSyncRuntime {
         pairingAttemptTimeoutNanoseconds: UInt64 = CMUXMobileRuntime.defaultPairingAttemptTimeoutNanoseconds,
         now: @escaping @Sendable () -> Date = Date.init,
         supportsServerPushEvents: Bool = true,
-        independentEventByteStreamProvider: CmxIndependentEventByteStreamProvider? = nil,
         terminalLaneProvider: MobileTerminalLaneProvider? = nil,
-        artifactLaneProvider: MobileArtifactLaneProvider? = nil,
         simulatorStreamLaneProvider: MobileSimulatorStreamLaneProvider? = nil
     ) {
         self.supportedRouteKinds = supportedRouteKinds
@@ -159,9 +156,7 @@ public struct CMUXMobileRuntime: Sendable, MobileSyncRuntime {
         self.pairingAttemptTimeoutNanoseconds = pairingAttemptTimeoutNanoseconds
         self.now = now
         self.supportsServerPushEvents = supportsServerPushEvents
-        self.independentEventByteStreamProvider = independentEventByteStreamProvider
         self.terminalLaneProvider = terminalLaneProvider
-        self.artifactLaneProvider = artifactLaneProvider
         self.simulatorStreamLaneProvider = simulatorStreamLaneProvider
     }
 
@@ -175,9 +170,7 @@ public struct CMUXMobileRuntime: Sendable, MobileSyncRuntime {
         pairingAttemptTimeoutNanoseconds: UInt64 = CMUXMobileRuntime.defaultPairingAttemptTimeoutNanoseconds,
         now: @escaping @Sendable () -> Date = Date.init,
         supportsServerPushEvents: Bool = true,
-        independentEventByteStreamProvider: CmxIndependentEventByteStreamProvider? = nil,
         terminalLaneProvider: MobileTerminalLaneProvider? = nil,
-        artifactLaneProvider: MobileArtifactLaneProvider? = nil,
         simulatorStreamLaneProvider: MobileSimulatorStreamLaneProvider? = nil
     ) {
         self.supportedRouteKinds = transportFactory.supportedKinds
@@ -189,9 +182,7 @@ public struct CMUXMobileRuntime: Sendable, MobileSyncRuntime {
         self.pairingRequestTimeoutNanoseconds = pairingRequestTimeoutNanoseconds
         self.pairingAttemptTimeoutNanoseconds = pairingAttemptTimeoutNanoseconds
         self.supportsServerPushEvents = supportsServerPushEvents
-        self.independentEventByteStreamProvider = independentEventByteStreamProvider
         self.terminalLaneProvider = terminalLaneProvider
-        self.artifactLaneProvider = artifactLaneProvider
         self.simulatorStreamLaneProvider = simulatorStreamLaneProvider
         self.now = now
     }
